Moved all related code to ODataResponseBuffer to a separate file, Renamed Test file to ODataDataObjectTest.scala
